Eliminate the need for the deprecated Prometheus CLI
This uses a combination of curl and jq instead.

Also, remove the 'in single quote' notes from the usage description,
because it is actually not true. Surely, you will set the single
quotes on the shell command line, but they are stripped off by the
shell. If you actually passed in single quotes (as in
`-H "'http://localhost'"`), things would break badly, even before
this change.